+++ src/api/uxout.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/skycoin/skycoin/src/cipher"
 	"github.com/skycoin/skycoin/src/readable"
-	wh "github.com/skycoin/skycoin/src/util/http" //http,json helpers
+	wh "github.com/skycoin/skycoin/src/util/http"
 )
 
 func getUxOutByID(gateway Gatewayer) http.HandlerFunc {